[new] Add Suika.writeImage()

+++ external/PlayfieldEngine/external/NoctLang/include/noct/noct.h
@@ -951,7 +951,7 @@ bool
 noct_get_array_elem_check_int(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	int *i);
 
@@ -977,7 +977,7 @@ bool
 noct_get_array_elem_check_long(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	int64_t *l);
 
@@ -1003,7 +1003,7 @@ bool
 noct_get_array_elem_check_float(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	float *f);
 
@@ -1029,7 +1029,7 @@ bool
 noct_get_array_elem_check_double(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	double *lf);
 
@@ -1050,7 +1050,7 @@ bool
 noct_get_array_elem_check_string(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	const char **data);
 
@@ -1071,7 +1071,7 @@ bool
 noct_get_array_elem_check_array(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val);
 
 /*
@@ -1091,7 +1091,7 @@ bool
 noct_get_array_elem_check_dict(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val);
 
 /*
@@ -1111,7 +1111,7 @@ bool
 noct_get_array_elem_check_func(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	NoctFunc **f);
 
@@ -1137,7 +1137,7 @@ bool
 noct_set_array_elem_make_int(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	int i);
 
@@ -1163,7 +1163,7 @@ bool
 noct_set_array_elem_make_long(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	int64_t l);
 
@@ -1189,7 +1189,7 @@ bool
 noct_set_array_elem_make_float(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	float f);
 
@@ -1215,7 +1215,7 @@ bool
 noct_set_array_elem_make_double(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	double lf);
 
@@ -1236,7 +1236,7 @@ bool
 noct_set_array_elem_make_string(
 	NoctEnv *env,
 	NoctValue *array,
-	uint32_t index,
+	size_t index,
 	NoctValue *val,
 	const char *data);
